requirements.txt
requests
pillow
twython
boto3